+++ apps/playground/src/app/(tabs)/import.tsx
@@ -18,7 +18,7 @@ import {
     extractAudioAnalysis,
     getWavFileInfo,
     SampleRate
-} from '@siteed/expo-audio-stream'
+} from '@siteed/expo-audio-studio'
 import { AudioVisualizer } from '@siteed/expo-audio-ui'
 import { Audio } from 'expo-av'
 import * as DocumentPicker from 'expo-document-picker'