+++ apps/playground/src/component/IOSSettingsConfig.tsx
@@ -1,5 +1,5 @@
 import { AppTheme, EditableInfoCard, useModal, useTheme } from '@siteed/design-system'
-import { RecordingConfig } from '@siteed/expo-audio-stream'
+import { RecordingConfig } from '@siteed/expo-audio-studio'
 import { getLogger } from '@siteed/react-native-logger'
 import React, { useState, useMemo } from 'react'
 import { Platform, StyleSheet, View, Text } from 'react-native'